requirements.txt
Pillow==9.3.0
moviepy==1.0.3